requirements.txt
aiogram>=3.13.0
Pillow>=10.0.0
mistralai>=1.0.0
google-api-python-client>=2.134.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
python-dotenv>=1.0.1
aiofiles>=23.2.1
